refactor: init all from index, move http clients and messaging modules
