+++ src/background/httpClients/httpClient.ts
@@ -1,0 +1,5 @@
+export const httpClient = {
+  fetch: async (url: string): Promise<Response> => {
+    return await fetch(url, { credentials: 'omit' });
+  },
+};
